WIP - Splited AppHeader into AppMenu and AppTitle.
* updated LICENSE year.
