+++ src/components/AppFooter.vue
@@ -1,13 +1,19 @@
 <template>
     <footer class="Footer">
-        <p><strong>Copyright&COPY; 2024 FlowerEvolver</strong></p><a href="https://github.com/cristianglezm/FlowerEvolver-frontend"><img alt="find-on-github" :src="data.base_url + 'find-on-github.png'"/></a>
+        <p>
+            <strong>Copyright&COPY; 2020-{{ data.currentYear }} FlowerEvolver</strong>
+        </p>
+        <a href="https://github.com/cristianglezm/FlowerEvolver-frontend">
+            <img alt="find-on-github" :src="data.base_url + 'find-on-github.png'"/>
+        </a>
     </footer>
 </template>
 
 <script setup>
 
 const data = {
-    base_url: import.meta.env.BASE_URL
+    base_url: import.meta.env.BASE_URL,
+    currentYear: new Date().getFullYear()
 }
 
 </script>
